Module Metrics_influx

val encode_line_protocol : Metrics.tags -> Metrics.data -> string -> string

encode_line_protocol tags data name encodes the tags and data to the influx line protocol, using name as measurement.

val lwt_reporter : ?⁠tags:Metrics.tags -> ?⁠interval:int -> (string -> unit Lwt.t) -> (unit -> int64) -> Metrics.reporter

lwt_reporter ~tags ~interval send clock is a metrics reporter that encodes a measurement in influxdb line protocol and reports it via send. A measurement can be prefixed by an optional list of tags. If ~interval is specified and a positive amount of milliseconds, each source measurement is reported only once within this interval.